Generate "DeviceName: SourceName" only with >1 source per device.
Pretty much everything was broken about this as it was previously implemented.  size was never 1 because it's in multiples of a UInt32's size (i.e., 4 on 32-bit systems).

Also fixes source name generation so it actually returns different things for each source name.  This was likely a result of confusion about the oddly named -fetchCoreAudioSourceType method, which returns the selected source type for the device ID (regardless of the source type the SSAudioDevice itself represents).