Always include “AirPlay: ” in AirPlay source description.
Otherwise, as the number of available AirPlay sources varies between one and more than one, the returned name changes, rendering scripts a bit useless.

We may very well need to similarly ensure that the device name and source name are always included for other transport types which can dynamically varying numbers of sources, but I’m not aware of any at the moment so will leave things as is.